+++ projects/js-packages/shared-extension-utils/index.js
@@ -4,3 +4,12 @@ export { isSimpleSite, isAtomicSite, isPrivateSite } from './src/site-type-utils
 export { default as getJetpackExtensionAvailability } from './src/get-jetpack-extension-availability';
 export { default as registerJetpackPlugin } from './src/register-jetpack-plugin';
 export { default as withHasWarningIsInteractiveClassNames } from './src/with-has-warning-is-interactive-class-names';
+export {
+	getUpgradeUrl,
+	isUpgradable,
+	requiresPaidPlan,
+	getRequiredPlan,
+	isUpgradeNudgeEnabled,
+	isStillUsableWithFreePlan,
+	getUsableBlockProps,
+} from './src/plan-utils';
