+++ projects/js-packages/shared-extension-utils/package.json
@@ -17,7 +17,10 @@
 	},
 	"dependencies": {
 		"@wordpress/compose": "5.1.2",
-		"@wordpress/plugins": "4.1.1"
+		"@wordpress/i18n": "4.3.0",
+		"@wordpress/plugins": "4.1.1",
+		"@wordpress/url": "3.4.0",
+		"lodash": "4.17.21"
 	},
 	"devDependencies": {
 		"@babel/core": "7.16.0",
