+++ projects/plugins/jetpack/extensions/extended-blocks/paid-blocks/media-replace-flow.js
@@ -1,6 +1,7 @@
 /**
  * External dependencies
  */
+import { isUpgradable, getUsableBlockProps } from '@automattic/jetpack-shared-extension-utils';
 
 /**
  * WordPress dependencies
@@ -13,7 +14,6 @@ import { useRef, useContext } from '@wordpress/element';
  * Internal dependencies
  */
 import { isFileOfType } from '../../shared/get-allowed-mime-types';
-import { isUpgradable, getUsableBlockProps } from '../../shared/plan-utils';
 import { PaidBlockContext } from './components';
 
 export default createHigherOrderComponent(
