+++ projects/plugins/jetpack/extensions/extended-blocks/paid-blocks/render-paid-icon.js
@@ -4,10 +4,14 @@
 import { cloneElement } from '@wordpress/element';
 
 /**
+ * External dependencies
+ */
+import { isUpgradable } from '@automattic/jetpack-shared-extension-utils';
+
+/**
  * Internal dependencies
  */
 import PaidSymbol from './paid-symbol';
-import { isUpgradable } from '../../shared/plan-utils';
 
 /**
  * Enhance the default block icon with a paid indicator
