+++ projects/plugins/jetpack/extensions/extended-blocks/paid-blocks/with-upgrade-banner.jsx
@@ -2,6 +2,11 @@
  * External dependencies
  */
 import classNames from 'classnames';
+import {
+	isStillUsableWithFreePlan,
+	getRequiredPlan,
+	getUsableBlockProps,
+} from '@automattic/jetpack-shared-extension-utils';
 
 /**
  * WordPress dependencies
@@ -13,11 +18,6 @@ import { useSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import {
-	isStillUsableWithFreePlan,
-	getRequiredPlan,
-	getUsableBlockProps,
-} from '../../shared/plan-utils';
 import UpgradePlanBanner from './upgrade-plan-banner';
 import { PaidBlockContext, PaidBlockProvider } from './components';
 import { trackUpgradeBannerImpression, trackUpgradeClickEvent } from './utils';
