+++ projects/plugins/jetpack/extensions/plugins/publicize/components/upsell/index.js
@@ -2,7 +2,11 @@
  * External dependencies
  */
 import classNames from 'classnames';
-import { isAtomicSite, isSimpleSite } from '@automattic/jetpack-shared-extension-utils';
+import {
+	isAtomicSite,
+	isSimpleSite,
+	getRequiredPlan,
+} from '@automattic/jetpack-shared-extension-utils';
 
 /**
  * WordPress dependencies
@@ -14,7 +18,6 @@ import { external } from '@wordpress/icons';
 /**
  * Internal dependencies
  */
-import { getRequiredPlan } from '../../../../shared/plan-utils';
 import useUpgradeFlow from '../../../../shared/use-upgrade-flow';
 import usePublicizeConfig from '../../hooks/use-publicize-config';
 import useSocialMediaConnections from '../../hooks/use-social-media-connections';
