+++ projects/plugins/jetpack/extensions/plugins/publicize/hooks/use-publicize-config/index.js
@@ -3,12 +3,10 @@
  */
 import { useSelect, useDispatch } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/editor';
-import { getJetpackExtensionAvailability } from '@automattic/jetpack-shared-extension-utils';
-
-/*
- * Internal dependencies
- */
-import { isUpgradable } from '../../../../shared/plan-utils';
+import {
+	getJetpackExtensionAvailability,
+	isUpgradable,
+} from '@automattic/jetpack-shared-extension-utils';
 
 const republicizeFeatureName = 'republicize';
