+++ projects/plugins/jetpack/extensions/shared/register-jetpack-block.js
@@ -7,6 +7,7 @@ import { registerBlockType } from '@wordpress/blocks';
 import {
 	getJetpackExtensionAvailability,
 	withHasWarningIsInteractiveClassNames,
+	requiresPaidPlan,
 } from '@automattic/jetpack-shared-extension-utils';
 
 /**
@@ -20,20 +21,6 @@ const availableBlockTags = {
 };
 
 const betaExtensions = extensionList.beta || [];
-
-/**
- * Checks whether the block requires a paid plan or not.
- *
- * @param {string} unavailableReason - The reason why block is unavailable
- * @param {object} details - The block details
- * @returns {string|boolean} Either false if the block doesn't require a paid plan, or the actual plan name it requires.
- */
-export function requiresPaidPlan( unavailableReason, details ) {
-	if ( unavailableReason === 'missing_plan' ) {
-		return details.required_plan;
-	}
-	return false;
-}
 
 /**
  * Builds an array of tags associated with this block, such as ["paid", "beta"].
