+++ projects/plugins/jetpack/extensions/shared/use-upgrade-flow/index.js
@@ -2,6 +2,7 @@
  * External dependencies
  */
 import { noop } from 'lodash';
+import { getUpgradeUrl } from '@automattic/jetpack-shared-extension-utils';
 
 /**
  * WordPress dependencies
@@ -15,7 +16,6 @@ import { doAction, hasAction } from '@wordpress/hooks';
 
 // Provably we should move this store to somewhere more generic.
 import '../components/upgrade-nudge/store';
-import { getUpgradeUrl } from '../plan-utils';
 import useAutosaveAndRedirect from '../use-autosave-and-redirect/index';
 
 const HOOK_OPEN_CHECKOUT_MODAL = 'a8c.wpcom-block-editor.openCheckoutModal';
